Using preserve_original option
It avoids need of translitarating words manually.

This strategy
does not enable actual accents-insensitive search, since
query text can contain original accents or no accents, but not wrong accents.
However, this is a less risky way of implementing this since we are not coupled
to strange rails i18n transliteration rules.