'Select one' fields in single-edit mode
The `notifyEdit` handler has been added in 2d7ba16676e8e731859a964af20ab48b54bec96b
to correctly activate single-edition mode while clicking on the
field's rows. But that handler was preventing the click event to
reach the `<select>` element, and thus avoided showing the options
- ie, it couldn't be edited.

Removing the listener seems to fix the Select one edition, and it
doesn't seem to break the row-click for any of the existent field
types.

Fixes #875