Empty commit to re-trigger TravisCI build
When we merged last commit, we were out of build credits in TravisCI,
so the build didn't even get generated. That means we cannot simply
rebuild the last attempt - there's no attempt to rebuild.

This new push will trigger a new build with the exact same code from
last commit.