Fix: compatibility with elasticsearch 2
Replaces multi_field type for a string with a nested field. The feature
was already deprecated in elasticsearch 1

Avoids the `.` in the field definition that is incompatible with
elasticsearch 2.

Removes the path definition that isn't needed and deprecated in
elasticsearch 2.

These changes are compatible with elasticsearch v1.7 and v2.4 and the
elasticsearch migration plugin returns a full green status.

See https://github.com/elastic/elasticsearch-migration/tree/1.x